move shape to storage
