+++ include/xtensor/xiterator.hpp
@@ -21,7 +21,6 @@
 
 #include "xexception.hpp"
 #include "xlayout.hpp"
-#include "xshape.hpp"
 #include "xutils.hpp"
 
 namespace xt
