+++ include/xtensor/xtensor_config.hpp
@@ -28,7 +28,7 @@
 
 #ifndef DEFAULT_SHAPE_CONTAINER
 #define DEFAULT_SHAPE_CONTAINER(T, EA, SA) \
-    xt::small_vector<typename DEFAULT_DATA_CONTAINER(T, EA)::size_type, 4, SA>
+    xt::svector<typename DEFAULT_DATA_CONTAINER(T, EA)::size_type, 4, SA>
 #endif
 
 #ifndef DEFAULT_ALLOCATOR
